requirements.txt
numpy>=1.24
scipy>=1.11
jax[cpu]>=0.4.30
pytest>=7.4
pyyaml>=6.0
torch>=2.1.0
pytorch-lightning>=2.2.0
tensorboard>=2.15.0
google-cloud-storage>=2.10.0
deepspeed>=0.13.0

pyarrow>=16.0.0
pydantic>=2.0.0
